pyproject.toml
[project]
name = "virtual-try-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1.1",
    "djangorestframework>=3.15.2",
    "opencv-python>=4.10.0.84",
    "mediapipe>=0.10.14",
    "numpy>=2.1.1",
    "Pillow>=11.0.0",
    "django-cors-headers>=4.4.0",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "python-decouple>=3.8",
]

